settings and figure2's picture
This is the settings for figure2. Also the picture is describing all experiments which are 8+3 totally. 8 is the model we need to train now, and 3 is the model we have trained before including atac-only, dna-only and atac+dna. And we finished the eighth experiment just now, whose pearson is close to zero indicating that our code is right.